refactor(query): isolate accepted coverage folding
